Updated select menu usage

+++ commands/private/vehicule.js
@@ -15,6 +15,8 @@ const regenVeh = require('../../modules/regenVehicles');
 
 const security = require('../../modules/service');
 
+let reply;
+
 module.exports = {
     //Création de la commande
     data: new SlashCommandBuilder()
@@ -40,8 +42,8 @@ module.exports = {
             .setRequired(true)
         ),
     async execute(interaction) {
-        const vehicules = await vehicles.get();
-        if(!security.isGen()) {
+        if(interaction.commandName == 'vehicule') {
+            const vehicules = await vehicles.get();
             if (interaction.options.getString(`action`) === `add`) {
                 if(hasAuthorization(Rank.DepartementManager, interaction.member.roles.cache)) {
                     const vehiculeAddModal = new ModalBuilder().setCustomId(`vehiculeAddModal`).setTitle(`Ajouter un véhicule`);
@@ -50,7 +52,6 @@ module.exports = {
                     const ct = new ActionRowBuilder().addComponents(new TextInputBuilder().setCustomId(`ct`).setLabel(`Date du contrôle technique`).setStyle(TextInputStyle.Short).setPlaceholder("Ex: JJ/MM/AAAA").setRequired(true));
                     const type = new ActionRowBuilder().addComponents(new TextInputBuilder().setCustomId(`type`).setLabel(`Type du véhicule`).setStyle(TextInputStyle.Short).setPlaceholder("Ex: fbi2").setRequired(true));
                     vehiculeAddModal.addComponents(nom, plaque, ct, type);
-                    security.setGen(true);
                     await interaction.showModal(vehiculeAddModal);
                 } else {
                     await interaction.reply({ embeds: [emb.generate(`Désolé :(`, null, `Vous n'avez pas les permissions suffisantes pour utiliser cette commande. Il faut être <@&${process.env.IRIS_DEPARTEMENT_MANAGER_ROLE}> ou plus pour pouvoir vous en servir !`, "#FF0000", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], ephemeral: true });
@@ -58,23 +59,30 @@ module.exports = {
                     await interaction.deleteReply();
                 }
             } else if(interaction.options.getString(`action`) === `regen`) {
-                //Affichage du message "Iris réfléchis..."
-                await interaction.deferReply({ ephemeral: true });
-                const allVeh = await vehicles.get();
-                const vehChanId = await vehicles.getChannelId();
-                if(vehChanId[0] == null) {
-                    return interaction.followUp({ embeds: [emb.generate(`Oups :(`, null, `Aucun salon de gestion des véhicules n'a été trouvé en base de donnée\nVeuillez contacter un de mes développeurs (<@461880599594926080>, <@461807010086780930> ou <@368259650136571904>) pour régler ce problème !`, "#FF0000", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], ephemeral: true });
+                if(!security.isGen()) {
+                    //Affichage du message "Iris réfléchis..."
+                    await interaction.deferReply({ ephemeral: true });
+                    const allVeh = await vehicles.get();
+                    const vehChanId = await vehicles.getChannelId();
+                    if(vehChanId[0] == null) {
+                        return interaction.followUp({ embeds: [emb.generate(`Oups :(`, null, `Aucun salon de gestion des véhicules n'a été trouvé en base de donnée\nVeuillez contacter un de mes développeurs (<@461880599594926080>, <@461807010086780930> ou <@368259650136571904>) pour régler ce problème !`, "#FF0000", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], ephemeral: true });
+                    }
+                    let vehiculeChannelID = vehChanId[0].id;
+                    const channelToSend = await interaction.guild.channels.cache.get(vehiculeChannelID);
+                    security.setGen(true);
+                    await regenVeh.all(channelToSend, allVeh);
+                    security.setGen(false);
+                    await interaction.followUp({ embeds: [emb.generate(null, null, `La liste des véhicules a bien été régénérée !`, "#0DE600", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], ephemeral: true });
+                    await wait(5000);
+                    await interaction.deleteReply();
+                } else {
+                    await interaction.reply({ embeds: [emb.generate(`Désolé :(`, null, `Désolé il y a déjà quelque chose en cours de génération, pour éviter tout problème veuillez patienter quelques instants puis réessayez !`, "#FF0000", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], ephemeral: true });
+                    await wait(5000);
+                    await interaction.deleteReply();
                 }
-                let vehiculeChannelID = vehChanId[0].id;
-                const channelToSend = await interaction.guild.channels.cache.get(vehiculeChannelID);
-                security.setGen(true);
-                await regenVeh.all(channelToSend, allVeh);
-                security.setGen(false);
-                await interaction.followUp({ embeds: [emb.generate(null, null, `La liste des véhicules a bien été régénérée !`, "#0DE600", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], ephemeral: true });
-                await wait(5000);
-                await interaction.deleteReply();
             } else if(interaction.options.getString(`action`) === `remove`) {
                 if(hasAuthorization(Rank.DepartementManager, interaction.member.roles.cache)) {
+                    reply = null;
                     const vehiculeRemoveSelect = new StringSelectMenuBuilder().setCustomId(`vehiculeRemoveSelect`).setPlaceholder(`Choisissez le(s) véhicule(s) à supprimer`).setMinValues(1);
                     //Ajout des opérations possibles
                     if(vehicules[0] == null) {
@@ -83,13 +91,12 @@ module.exports = {
                         await interaction.deleteReply();
                     } else {
                         //Affichage du message "Iris réfléchis..."
-                        await interaction.deferReply({ ephemeral: true });
+                        reply = await interaction.deferReply({ ephemeral: true });
                         for(i=0;i<vehicules.length;i++) {
                             vehiculeRemoveSelect.addOptions(new StringSelectMenuOptionBuilder().setLabel(`${vehicules[i].name}`).setValue(`${vehicules[i].plate}`));
                         }
                         vehiculeRemoveSelect.setMaxValues(vehicules.length);
                         const allOptions = new ActionRowBuilder().addComponents(vehiculeRemoveSelect);
-                        security.setGen(true);
                         await interaction.followUp({ components: [allOptions], ephemeral: true });
                     }
                 } else {
@@ -98,10 +105,80 @@ module.exports = {
                     await interaction.deleteReply();
                 }
             }
-        } else {
-            await interaction.reply({ embeds: [emb.generate(`Désolé :(`, null, `Désolé il y a déjà quelque chose en cours de génération, pour éviter tout problème veuillez patienter quelques instants puis réessayez !`, "#FF0000", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], ephemeral: true });
-            await wait(5000);
-            await interaction.deleteReply();
+        } else if(interaction.customId == 'vehiculeRemoveSelect') {
+            if(!security.isGen()) {
+                let respContent = 'Le(s) véhicule(s) immatriculé(s) ';
+                let isVehNull = false;
+                for(i=0;i<interaction.values.length;i++) {
+                    const vehicules = await vehicles.getByPlate(interaction.values[i]);
+                    if(vehicules[0] != null) {
+                        await vehicles.delete(vehicules[0].plate);
+                        if(i == 0) {
+                            respContent = respContent + `**${vehicules[0].plate}**`
+                        } else if(i != interaction.values.length - 1) {
+                            respContent = respContent + `, **${vehicules[0].plate}**`
+                        } else {
+                            respContent = respContent + ` et **${vehicules[0].plate}**`
+                        }
+                    } else {
+                        isVehNull = true;
+                    }
+                }
+                //Send confirmation message
+                if(!isVehNull) {
+                    const allVehicles = await vehicles.get();
+                    //Récupération du channel des véhicules
+                    const vehChanId = await vehicles.getChannelId();
+                    try {
+                        if(vehChanId[0] == null) {
+                            return reply.edit({ embeds: [emb.generate(`Oups :(`, null, `Aucun salon de gestion des véhicules n'a été trouvé en base de donnée\nVeuillez contacter un des développeurs (<@461880599594926080>, <@461807010086780930> ou <@368259650136571904>) pour régler ce problème !`, "#FF0000", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], components: [], ephemeral: true });
+                        }
+                        await reply.edit({ embeds: [emb.generate(null, null, `<a:loading:1137771560531398697> Mise à jour en cours...`, `Gold`, process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], components: [], ephemeral: true });
+                        let vehiculeChannelID = vehChanId[0].id;
+                        const vehChan = await interaction.guild.channels.cache.get(vehiculeChannelID);
+                        security.setGen(true);
+                        await regenVeh.all(vehChan, allVehicles);
+                        security.setGen(false);
+                        await reply.edit({ embeds: [emb.generate(null, null, `${respContent} a/ont bien été retiré(s) de la liste !`, `#0DE600`, process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], components: [], ephemeral: true });
+                        // Supprime la réponse après 5s
+                        await wait(5000);
+                        await reply.delete();
+                    } catch (err) {
+                        logger.error(err);
+                        if(vehChanId[0] == null) {
+                            return interaction.reply({ embeds: [emb.generate(`Oups :(`, null, `Aucun salon de gestion des véhicules n'a été trouvé en base de donnée\nVeuillez contacter un des développeurs (<@461880599594926080>, <@461807010086780930> ou <@368259650136571904>) pour régler ce problème !`, "#FF0000", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], components: [], ephemeral: true });
+                        }
+                        interaction.deferUpdate();
+                        let vehiculeChannelID = vehChanId[0].id;
+                        const vehChan = await interaction.guild.channels.cache.get(vehiculeChannelID);
+                        security.setGen(true);
+                        await regenVeh.all(vehChan, allVehicles);
+                        security.setGen(false);
+                        await interaction.reply({ embeds: [emb.generate(null, null, `${respContent} a/ont bien été retiré(s) de la liste !`, `#0DE600`, process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], ephemeral: true });
+                        // Supprime la réponse après 5s
+                        await wait(5000);
+                        await interaction.deleteReply();
+                    }
+                } else {
+                    try {
+                        await reply.edit({ embeds: [emb.generate(null, null, `⚠️ Attention, il n'y a plus aucun véhicule encore présent dans cette liste à supprimer\nSi jamais il en reste encore veuillez refaire la commande !`, `Gold`, process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], components: [], ephemeral: true });
+                    } catch (err) {
+                        logger.error(err);
+                        await interaction.reply({ embeds: [emb.generate(null, null, `⚠️ Attention, il n'y a plus aucun véhicule encore présent dans cette liste à supprimer\nSi jamais il en reste encore veuillez refaire la commande !`, `Gold`, process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], ephemeral: true });
+                    }
+                }
+            } else {
+                try {
+                    await reply.edit({ embeds: [emb.generate(`Désolé :(`, null, `Désolé il y a déjà quelque chose en cours de génération, pour éviter tout problème veuillez patienter quelques instants puis réessayez !`, "#FF0000", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], components: [], ephemeral: true });
+                    await wait(5000);
+                    await reply.delete();
+                } catch (err) {
+                    logger.error(err);
+                    await interaction.reply({ embeds: [emb.generate(`Désolé :(`, null, `Désolé il y a déjà quelque chose en cours de génération, pour éviter tout problème veuillez patienter quelques instants puis réessayez !`, "#FF0000", process.env.LSMS_LOGO_V2, null, `Gestion des véhicules`, `https://cdn.discordapp.com/icons/${process.env.IRIS_PRIVATE_GUILD_ID}/${interaction.guild.icon}.webp`, null, null, null, false)], components: [], ephemeral: true });
+                    await wait(5000);
+                    await interaction.deleteReply();
+                }
+            }
         }
     },
 };
